fix module.exports in index.js and move index.js contents to get-link.js
fix module.exports in index.js and move index.js contents to get-link.js
