+++ get-link.js
@@ -1,0 +1,3 @@
+const gitRemoteToUrl = require('./git-remote-to-url');
+
+console.log(gitRemoteToUrl(process.argv[2]));
