+++ index.js
@@ -1,3 +1,1 @@
-const gitRemoteToUrl = require('./gitRemoteToUrl');
-
-console.log(gitRemoteToUrl(process.argv[2]));
+module.exports = require('./git-remote-to-url');
